requirements.txt
Django==3.0.8
Jinja2==2.11.2
MarkupSafe==1.1.1
Pillow==6.2.2
Pygments==2.7.3
asgiref==3.2.10
certifi==2020.12.5
chardet==3.0.4
django-better-admin-arrayfield==1.4.2
django-cors-headers==3.6.0
django-filter==2.4.0
django-jsonfield==1.4.1
django-multiselectfield==0.1.12
djangorestframework==3.12.2
et-xmlfile==1.0.1
haversine==2.3.0
idna==2.10
install==1.3.4
jdcal==1.4.1
mysqlclient==2.0.2
numpy==1.19.3
openpyxl==3.0.5
pip==20.3.3
psycopg2==2.8.6
pytz==2020.1
requests==2.24.0
setuptools==39.0.1
six	1.15.0==1.15.0
sqlparse==0.3.1
urllib3==1.25.11